+++ guillotina/factory/security.py
@@ -2,10 +2,13 @@ from guillotina import configure
 from guillotina import security
 from guillotina.auth.users import ANONYMOUS_USER_ID
 from guillotina.auth.users import ROOT_USER_ID
+from guillotina.interfaces import AllowSingle
 from guillotina.interfaces import IApplication
 from guillotina.interfaces import IDatabase
 from guillotina.interfaces import IPrincipalPermissionManager
+from guillotina.interfaces import IRolePermissionManager
 from guillotina.security.security_code import PrincipalPermissionManager
+from guillotina.security.security_code import RolePermissionManager
 
 
 @configure.adapter(for_=IDatabase, provides=IPrincipalPermissionManager)
@@ -59,8 +62,17 @@ class ApplicationSpecialPermissions(DatabaseSpecialPermissions):
             # Access anonymous - needs to be configurable
             # so anon can access static content mostly
             self.grant_permission_to_principal(
-                'guillotina.AccessContent', ANONYMOUS_USER_ID)
+                'guillotina.AccessContent', ANONYMOUS_USER_ID, mode=AllowSingle)
             app.__cached_map__ = {
                 'byrow': self._byrow,
                 'bycol': self._bycol
             }
+
+@configure.adapter(for_=IApplication, provides=IRolePermissionManager)
+class ApplicationSpecialRoles(RolePermissionManager):
+    def __init__(self, app):
+        super().__init__()
+        self.grant_permission_to_role(
+            'guillotina.AccessContent', 'guillotina.Authenticated',
+            mode=AllowSingle
+        )